+++ R/compound_taskid-utils.R
@@ -12,9 +12,28 @@
 #' @return A list of vectors of compound task IDs detected in the tbl, one for each
 #' modeling task in the round. If `compact` is TRUE, modeling tasks returning NULL
 #' elements will be removed.
-#' @noRd
+#' @export
+#' @examples
+#' hub_path <- system.file("testhubs/samples", package = "hubValidations")
+#' file_path <- "Flusight-baseline/2022-10-22-Flusight-baseline.csv"
+#' round_id <- "2022-10-22"
+#' tbl <- read_model_out_file(
+#'   file_path = file_path,
+#'   hub_path = hub_path,
+#'   coerce_types = "chr"
+#' )
+#' config_tasks <- hubUtils::read_config(hub_path, "tasks")
+#' get_tbl_compound_taskid_set(tbl, config_tasks, round_id)
+#' get_tbl_compound_taskid_set(tbl, config_tasks, round_id,
+#'   compact = FALSE
+#' )
+#'
 get_tbl_compound_taskid_set <- function(tbl, config_tasks, round_id,
                                         compact = TRUE, error = TRUE) {
+  if (!inherits(tbl, "tbl_df")) {
+    tbl <- dplyr::as_tibble(tbl)
+  }
+
   tbl <- tbl[tbl$output_type == "sample", names(tbl) != "value"]
   out_tid <- hubUtils::std_colnames["output_type_id"]
 
@@ -174,7 +193,7 @@ get_mt_compound_taskid_set <- function(tbl, config_comp_tids, config_tasks,
 # Extract the names of the columns that are TRUE in each row
 true_to_names_vector <- function(x, cols = NULL, unique = TRUE) {
   if (!is.null(cols)) {
-    x <- x[, cols]
+    x <- x[, cols, drop = FALSE]
   }
   if (unique) {
     x <- unique(x)